lots stuff, mainly panning & singleton
done:
- sorted console key suggestions
- alert on class.interface = true when had instances
- singleton beta
- final (incomplete)

new css variables:
--offset-x (panning del grafo)
--offset-y
--zoom-x (del grafo, ma lo zoom non funziona ancora)
--zoom-y
--top (node.y)
--left (node.x)

new features engine:
- aggregated (class/object)
- composed (class/object)
- contained (class/object) = aggregated || composed
PS: aggregation and composition exist in UML only, containment in eCore only, i tried to cover both.
Those properties are "is class/object pointed by a X-type reference?"
- instantiable (class) = !abstract && !interface && !singleton // just a shortcut
- rootable (class) = instantiable && !contained // just a shortcut
- referencedBy (class/object) = array of LReference/LValue pointing here

bugfixes:
primitive arrays showing as objects in console
color palette picker cropping & similar layout stuff

+++ src/api/persistance/auth.ts
@@ -12,7 +12,7 @@ class AuthApi {
         return await Api.post(`${Api.persistance}/auth/register`, {username, email, password});
     }
     static async logout(): Promise<void> {
-        await Api.delete(`${Api.persistance}/auth/logout`);
+        if(!U.isOffline()) await Api.delete(`${Api.persistance}/auth/logout`);
         Storage.reset();
         U.refresh();
     }
